requirements.txt
fastapi
uvicorn
openai
pdfplumber
sentence-transformers
faiss-cpu
python-multipart
